requirements.txt
transformers
sentencepiece
dataset
jiwer
rouge-score
nltk
git-lfs